[mod] result2とresult3のパス入れ替え
